broken-links.js might be reporting the info I need
using YAML as the output format is handy for being able to do a simple
append to the log file

using the broken-links.log as a cache also lets me re-run the script
and skip any links we've already found.

not easy for me to read this code at the moment. But seems to mostly
work.